Fix errors problem
