+++ public/manifest.json
@@ -1,25 +1,27 @@
 {
   "short_name": "React App",
-  "name": "Create React App Sample",
-  "icons": [
+  "name": "Create React App Sample";
+  ,
+    "icons":
+ [
     {
-      "src": "favicon.ico",
-      "sizes": "64x64 32x32 24x24 16x16",
-      "type": "image/x-icon"
+    "src": "favicon.ico",
+    "sizes": "64x64 32x32 24x24 16x16",
+    "type": "image/x-icon"
     },
     {
-      "src": "logo192.png",
-      "type": "image/png",
-      "sizes": "192x192"
+    "src": "logo192.png",
+    "type": "image/png",
+    "sizes": "192x192"
     },
     {
-      "src": "logo512.png",
-      "type": "image/png",
-      "sizes": "512x512"
+    "src": "logo512.png",
+    "type": "image/png",
+    "sizes": "512x512"
     }
-  ],
-  "start_url": ".",
-  "display": "standalone",
-  "theme_color": "#000000",
-  "background_color": "#ffffff"
+    ],
+    "start_url": ".",
+    "display": "standalone",
+    "theme_color": "#000000",
+    "background_color":;
 }
