+++ src/components/App.css
@@ -1,0 +1,3 @@
+body {
+  color: rgb(204, 203, 203);
+}
