+++ src/components/Greeting.js
@@ -1,0 +1,23 @@
+import React, { useEffect } from 'react';
+import { useSelector, useDispatch } from 'react-redux';
+import { fetchGreeting } from './redux/greeting';
+
+const Greeting = () => {
+  const dispatch = useDispatch();
+  const greeting = useSelector((state) => state.greeting);
+
+  useEffect(() => {
+    dispatch(fetchGreeting());
+  }, []);
+
+  return (
+    <div>
+      <h1>
+        Greeting:
+        {greeting.message}
+      </h1>
+    </div>
+  );
+};
+
+export default Greeting;
