+++ src/components/redux/configureStore.js
@@ -1,0 +1,11 @@
+import { createStore, combineReducers, applyMiddleware } from 'redux';
+import thunk from 'redux-thunk';
+import logger from 'redux-logger';
+import greetingReducer from './greeting';
+
+const reducer = combineReducers({
+  greeting: greetingReducer,
+});
+
+const store = createStore(reducer, {}, applyMiddleware(thunk, logger));
+export default store;
